Move GCP Container to core
