+++ airflow/contrib/hooks/gcp_container_hook.py
@@ -16,271 +16,16 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-#
 """
-This module contains a Google Kubernetes Engine Hook.
+This module is deprecated. Please use `airflow.gcp.hooks.container`.
 """
 
-import json
-import time
-from typing import Dict, Union, Optional
+import warnings
 
-from google.api_core.exceptions import AlreadyExists, NotFound
-from google.api_core.gapic_v1.method import DEFAULT
-from google.api_core.retry import Retry
+# pylint: disable=unused-import
+from airflow.gcp.hooks.container import GKEClusterHook  # noqa
 
-from google.cloud import container_v1, exceptions
-from google.cloud.container_v1.gapic.enums import Operation
-from google.cloud.container_v1.types import Cluster
-from google.protobuf import json_format
-
-from airflow import AirflowException, version
-from airflow.contrib.hooks.gcp_api_base_hook import GoogleCloudBaseHook
-
-OPERATIONAL_POLL_INTERVAL = 15
-
-
-class GKEClusterHook(GoogleCloudBaseHook):
-    """
-    Hook for Google Kubernetes Engine APIs.
-
-    All the methods in the hook where project_id is used must be called with
-    keyword arguments rather than positional.
-    """
-
-    def __init__(
-        self,
-        gcp_conn_id: str = 'google_cloud_default',
-        delegate_to: str = None,
-        location: str = None
-    ) -> None:
-        super().__init__(
-            gcp_conn_id=gcp_conn_id, delegate_to=delegate_to)
-        self._client = None
-        self.location = location
-
-    def get_client(self) -> container_v1.ClusterManagerClient:
-        """
-        Returns ClusterManagerCLinet object.
-
-        :rtype: google.cloud.container_v1.ClusterManagerClient
-        """
-        if self._client is None:
-            credentials = self._get_credentials()
-            self._client = container_v1.ClusterManagerClient(
-                credentials=credentials,
-                client_info=self.client_info
-            )
-        return self._client
-
-    @staticmethod
-    def _dict_to_proto(py_dict: Dict, proto):
-        """
-        Converts a python dictionary to the proto supplied
-
-        :param py_dict: The dictionary to convert
-        :type py_dict: dict
-        :param proto: The proto object to merge with dictionary
-        :type proto: protobuf
-        :return: A parsed python dictionary in provided proto format
-        :raises:
-            ParseError: On JSON parsing problems.
-        """
-        dict_json_str = json.dumps(py_dict)
-        return json_format.Parse(dict_json_str, proto)
-
-    def wait_for_operation(self, operation: Operation, project_id: str = None) -> Operation:
-        """
-        Given an operation, continuously fetches the status from Google Cloud until either
-        completion or an error occurring
-
-        :param operation: The Operation to wait for
-        :type operation: google.cloud.container_V1.gapic.enums.Operation
-        :param project_id: Google Cloud Platform project ID
-        :type project_id: str
-        :return: A new, updated operation fetched from Google Cloud
-        """
-        self.log.info("Waiting for OPERATION_NAME %s", operation.name)
-        time.sleep(OPERATIONAL_POLL_INTERVAL)
-        while operation.status != Operation.Status.DONE:
-            if operation.status == Operation.Status.RUNNING or operation.status == \
-                    Operation.Status.PENDING:
-                time.sleep(OPERATIONAL_POLL_INTERVAL)
-            else:
-                raise exceptions.GoogleCloudError(
-                    "Operation has failed with status: %s" % operation.status)
-            # To update status of operation
-            operation = self.get_operation(operation.name, project_id=project_id or self.project_id)
-        return operation
-
-    def get_operation(self, operation_name: str, project_id: str = None) -> Operation:
-        """
-        Fetches the operation from Google Cloud
-
-        :param operation_name: Name of operation to fetch
-        :type operation_name: str
-        :param project_id: Google Cloud Platform project ID
-        :type project_id: str
-        :return: The new, updated operation from Google Cloud
-        """
-        return self.get_client().get_operation(project_id=project_id or self.project_id,
-                                               zone=self.location,
-                                               operation_id=operation_name)
-
-    @staticmethod
-    def _append_label(cluster_proto: Cluster, key: str, val: str) -> Cluster:
-        """
-        Append labels to provided Cluster Protobuf
-
-        Labels must fit the regex ``[a-z]([-a-z0-9]*[a-z0-9])?`` (current
-         airflow version string follows semantic versioning spec: x.y.z).
-
-        :param cluster_proto: The proto to append resource_label airflow
-            version to
-        :type cluster_proto: google.cloud.container_v1.types.Cluster
-        :param key: The key label
-        :type key: str
-        :param val:
-        :type val: str
-        :return: The cluster proto updated with new label
-        """
-        val = val.replace('.', '-').replace('+', '-')
-        cluster_proto.resource_labels.update({key: val})
-        return cluster_proto
-
-    def delete_cluster(
-        self,
-        name: str,
-        project_id: str = None,
-        retry: Retry = DEFAULT,
-        timeout: float = DEFAULT
-    ) -> Optional[str]:
-        """
-        Deletes the cluster, including the Kubernetes endpoint and all
-        worker nodes. Firewalls and routes that were configured during
-        cluster creation are also deleted. Other Google Compute Engine
-        resources that might be in use by the cluster (e.g. load balancer
-        resources) will not be deleted if they weren’t present at the
-        initial create time.
-
-        :param name: The name of the cluster to delete
-        :type name: str
-        :param project_id: Google Cloud Platform project ID
-        :type project_id: str
-        :param retry: Retry object used to determine when/if to retry requests.
-            If None is specified, requests will not be retried.
-        :type retry: google.api_core.retry.Retry
-        :param timeout: The amount of time, in seconds, to wait for the request to
-            complete. Note that if retry is specified, the timeout applies to each
-            individual attempt.
-        :type timeout: float
-        :return: The full url to the delete operation if successful, else None
-        """
-
-        self.log.info(
-            "Deleting (project_id=%s, zone=%s, cluster_id=%s)", self.project_id, self.location, name
-        )
-
-        try:
-            resource = self.get_client().delete_cluster(project_id=project_id or self.project_id,
-                                                        zone=self.location,
-                                                        cluster_id=name,
-                                                        retry=retry,
-                                                        timeout=timeout)
-            resource = self.wait_for_operation(resource)
-            # Returns server-defined url for the resource
-            return resource.self_link
-        except NotFound as error:
-            self.log.info('Assuming Success: %s', error.message)
-            return None
-
-    def create_cluster(
-        self,
-        cluster: Union[Dict, Cluster],
-        project_id: str = None,
-        retry: Retry = DEFAULT,
-        timeout: float = DEFAULT
-    ) -> str:
-        """
-        Creates a cluster, consisting of the specified number and type of Google Compute
-        Engine instances.
-
-        :param cluster: A Cluster protobuf or dict. If dict is provided, it must
-            be of the same form as the protobuf message
-            :class:`google.cloud.container_v1.types.Cluster`
-        :type cluster: dict or google.cloud.container_v1.types.Cluster
-        :param project_id: Google Cloud Platform project ID
-        :type project_id: str
-        :param retry: A retry object (``google.api_core.retry.Retry``) used to
-            retry requests.
-            If None is specified, requests will not be retried.
-        :type retry: google.api_core.retry.Retry
-        :param timeout: The amount of time, in seconds, to wait for the request to
-            complete. Note that if retry is specified, the timeout applies to each
-            individual attempt.
-        :type timeout: float
-        :return: The full url to the new, or existing, cluster
-        :raises:
-            ParseError: On JSON parsing problems when trying to convert dict
-            AirflowException: cluster is not dict type nor Cluster proto type
-        """
-
-        if isinstance(cluster, dict):
-            cluster_proto = Cluster()
-            cluster = self._dict_to_proto(py_dict=cluster, proto=cluster_proto)
-        elif not isinstance(cluster, Cluster):
-            raise AirflowException(
-                "cluster is not instance of Cluster proto or python dict")
-
-        self._append_label(cluster, 'airflow-version', 'v' + version.version)
-
-        self.log.info(
-            "Creating (project_id=%s, zone=%s, cluster_name=%s)",
-            self.project_id, self.location, cluster.name
-        )
-        try:
-            resource = self.get_client().create_cluster(project_id=project_id or self.project_id,
-                                                        zone=self.location,
-                                                        cluster=cluster,
-                                                        retry=retry,
-                                                        timeout=timeout)
-            resource = self.wait_for_operation(resource)
-
-            return resource.target_link
-        except AlreadyExists as error:
-            self.log.info('Assuming Success: %s', error.message)
-            return self.get_cluster(name=cluster.name).self_link
-
-    def get_cluster(
-        self,
-        name: str,
-        project_id: str = None,
-        retry: Retry = DEFAULT,
-        timeout: float = DEFAULT
-    ) -> Cluster:
-        """
-        Gets details of specified cluster
-
-        :param name: The name of the cluster to retrieve
-        :type name: str
-        :param project_id: Google Cloud Platform project ID
-        :type project_id: str
-        :param retry: A retry object used to retry requests. If None is specified,
-            requests will not be retried.
-        :type retry: google.api_core.retry.Retry
-        :param timeout: The amount of time, in seconds, to wait for the request to
-            complete. Note that if retry is specified, the timeout applies to each
-            individual attempt.
-        :type timeout: float
-        :return: google.cloud.container_v1.types.Cluster
-        """
-        self.log.info(
-            "Fetching cluster (project_id=%s, zone=%s, cluster_name=%s)",
-            project_id or self.project_id, self.location, name
-        )
-
-        return self.get_client().get_cluster(project_id=project_id or self.project_id,
-                                             zone=self.location,
-                                             cluster_id=name,
-                                             retry=retry,
-                                             timeout=timeout).self_link
+warnings.warn(
+    "This module is deprecated. Please use `airflow.gcp.hooks.container`",
+    DeprecationWarning,
+)
