+++ docs/integration.rst
@@ -745,16 +745,16 @@ They also use :class:`airflow.gcp.hooks.video_intelligence.CloudVideoIntelligenc
 Google Kubernetes Engine
 ''''''''''''''''''''''''
 
-:class:`airflow.contrib.operators.gcp_container_operator.GKEClusterCreateOperator`
+:class:`airflow.gcp.operators.container.GKEClusterCreateOperator`
     Creates a Kubernetes Cluster in Google Cloud Platform
 
-:class:`airflow.contrib.operators.gcp_container_operator.GKEClusterDeleteOperator`
+:class:`airflow.gcp.operators.container.GKEClusterDeleteOperator`
     Deletes a Kubernetes Cluster in Google Cloud Platform
 
-:class:`airflow.contrib.operators.gcp_container_operator.GKEPodOperator`
+:class:`airflow.gcp.operators.container.GKEPodOperator`
     Executes a task in a Kubernetes pod in the specified Google Kubernetes Engine cluster
 
-They also use :class:`airflow.contrib.hooks.gcp_container_hook.GKEClusterHook` to communicate with Google Cloud Platform.
+They also use :class:`airflow.gcp.hooks.container.GKEClusterHook` to communicate with Google Cloud Platform.
 
 
 Google Natural Language
